Fix different warnings from previous commits.

+++ src/gui/widgets/namesmodel.cpp
@@ -52,6 +52,6 @@ std::string NamesModel::getElementAt(int i)
 
 void NamesModel::fillFromArray(const char *const *const arr, std::size_t sz)
 {
-    for (int f = 0; f < sz; f ++)
+    for (size_t f = 0; f < sz; f ++)
         mNames.push_back(gettext(arr[f]));
 }
